Move shared mock_msg_queue to it's own file
